requirements.txt
accelerate==1.2.1
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anthropic==0.31.0
anyio==4.4.0
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
audioread==3.0.1
av==14.0.1
bitsandbytes==0.43.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.0.0
cmake==3.30.0
compressed-tensors==0.6.0
contourpy==1.2.1
cycler==0.12.1
datasets==2.20.0
decorator==5.1.1
decord==0.6.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
easydict==1.13
einops==0.6.1
einops-exts==0.0.4
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
gguf==0.10.0
gradio==4.16.0
gradio_client==0.8.1
h11==0.14.0
hf_transfer==0.1.6
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.0
huggingface-hub==0.27.0
idna==3.10
importlib_metadata==8.0.0
importlib_resources==6.4.0
interegular==0.3.3
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lark==1.1.9
latex2mathml==3.77.0
lazy_loader==0.4
librosa==0.10.2.post1
litellm==1.41.19
llvmlite==0.43.0
lm-format-enforcer==0.10.6
markdown-it-py==3.0.0
markdown2==2.5.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mistral_common==1.5.1
modelscope==1.21.0
mpmath==1.3.0
msgpack==1.0.8
msgspec==0.18.6
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.555.43
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
openai==1.50.2
opencv-python-headless==4.10.0.84
orjson==3.10.6
outlines==0.0.46
packaging==24.1
pandas==2.2.2
partial-json-parser==0.2.1.1.post4
peft==0.11.1
pillow==10.4.0
platformdirs==4.3.6
plumbum==1.8.3
pooch==1.8.2
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
protobuf==5.27.2
psutil==6.0.0
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==16.1.0
pyarrow-hotfix==0.6
pycountry==24.6.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
qwen-vl-utils==0.0.8
ray==2.32.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
rpyc==6.0.0
ruff==0.5.1
safetensors==0.4.3
scikit-learn==1.2.2
scipy==1.14.0
semantic-version==2.10.0
sentencepiece==0.2.0
sglang==0.3.4.post2
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
starlette==0.37.2
svgwrite==1.4.3
sympy==1.13.0
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==0.6.13
tokenizers==0.20.3
tomlkit==0.12.0
toolz==0.12.1
torch==2.4.0
torchao==0.5.0
torchvision==0.19.0
tqdm==4.66.5
transformers==4.46.0
triton==3.0.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.30.1
uvloop==0.19.0
vllm==0.6.3.post1
vllm-flash-attn==2.6.1
watchfiles==0.22.0
wavedrom==2.0.3.post3
websockets==11.0.3
xformers==0.0.27.post2
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
zmq==0.0.0
